+++ src/ast/expressions/function/function.spec.ts
@@ -22,7 +22,7 @@ const fn = (name: string, ...args: Expression[]) =>
       None(),
       Right(ValueType.Boolean)));
 
-const phrase = (val: string): PhraseExpression<any> =>
+const phrase = (val: string): PhraseExpression =>
   PhraseExpression.fromTerm(fromMatch(config)(val));
 
 describe('SearchQL expressions', () => {
@@ -224,7 +224,6 @@ describe('SearchQL expressions', () => {
           returnType: ValueType.Boolean,
           type: ExpressionType.Function,
           value: [{
-            preparedValue: 'lorem',
             returnType: ValueType.Text,
             type: ExpressionType.Text,
             value: 'lorem',
@@ -246,8 +245,12 @@ describe('SearchQL expressions', () => {
           returnType: ValueType.Boolean,
           type: ExpressionType.Function,
           value: [{
-            preparedValue: 1514764800000,
             returnType: ValueType.Date,
+            timeFrame: {
+              end: 1514879999999,
+              start: 1514793600000,
+              text: '2018-01-01',
+            },
             type: ExpressionType.Date,
             value: '2018-01-01',
           }],