+++ src/ast/expressions/term/is-date.ts
@@ -1,1 +1,0 @@
-export const isDate = (value: string) => new Date(value.trim()).toJSON() != null;