+++ src/ast/expressions/term/selector.ts
@@ -7,7 +7,7 @@ import { InvalidExpression } from '../invalid';
 import { PhraseExpression } from './phrase';
 import { TermExpression } from './term';
 
-export class SelectorExpression extends TermExpression<string> {
+export class SelectorExpression extends TermExpression {
 
   public static of(value: string) {
     return (matchingType: ValueType) =>
@@ -19,15 +19,13 @@ export class SelectorExpression extends TermExpression<string> {
   constructor(
     public readonly matchingType: ValueType,
     value: string,
-    preparedValue = value.trim(),
   ) {
-    super(value, preparedValue);
+    super(value);
   }
 
   public toJS(): ISelectorExpression {
     return {
       matchingType: this.matchingType,
-      preparedValue: this.preparedValue,
       returnType: this.returnType,
       type: this.type,
       value: this.value,
@@ -48,30 +46,22 @@ export class SelectorExpression extends TermExpression<string> {
     return Some([
       ...this.getNotInModelError(model),
       ...this.getNotMatchingModelTypeError(model),
-      ...this.getNotMatchingValuesError(),
     ]).filter(errors => errors.length > 0);
   }
 
   private getNotInModelError(model: Map<string, ValueType>) {
-    return model.has(this.preparedValue) ? [] : [
-      `SelectorExpression is invalid - "${this.preparedValue}" is ` +
+    return model.has(this.value) ? [] : [
+      `SelectorExpression is invalid - "${this.value}" is ` +
         'not available on model definition',
     ];
   }
 
   private getNotMatchingModelTypeError(model: Map<string, ValueType>) {
-    return model.get(this.preparedValue) === this.matchingType ? [] :
-      model.has(this.preparedValue) ? [
+    return model.get(this.value) === this.matchingType ? [] :
+      model.has(this.value) ? [
         'SelectorExpression is invalid - "matchingType" should equal ' +
-          `${model.get(this.preparedValue)} but is ${this.matchingType}`,
+          `${model.get(this.value)} but is ${this.matchingType}`,
       ] : [];
-  }
-
-  private getNotMatchingValuesError() {
-    return this.value.trim() === this.preparedValue ? [] : [
-      'Values of SelectorExpression don\'t match: ' +
-        `{ value: ${this.value.trim()}, preparedValue: ${this.preparedValue}}`,
-    ];
   }
 
 }