+++ src/ast/from-js.ts
@@ -1,7 +1,7 @@
 import { List } from 'immutable';
-import { Expression, ExpressionType } from '../common/model';
+import { Expression, ExpressionType, TimeFrame } from '../common/model';
 import { FunctionConfig } from '../config';
-import { IBinaryOperationExpression, IExpression, IFunctionExpression, INotExpression, IPhraseExpression, ISelectorExpression, ITermExpression } from '../dto';
+import { IBinaryOperationExpression, IDateExpression, IExpression, IFunctionExpression, INotExpression, IPhraseExpression, ISelectorExpression, ITermExpression } from '../dto';
 import { BinaryOperationExpression, DateExpression, FunctionExpression, NotExpression, NumberExpression, PhraseExpression, SelectorExpression, TermExpression, TextExpression } from './expressions';
 import { Operator } from './operators';
 
@@ -10,27 +10,27 @@ export const fromJS = <E extends IExpression>(pojo: E): Expression => {
 
   switch (pojo.type) {
     case ExpressionType.Date: {
-      const { preparedValue, value } = (pojo as any as ITermExpression<number>);
-      return new DateExpression(value, preparedValue);
+      const { timeFrame, value } = (pojo as any as IDateExpression);
+      return new DateExpression(value, TimeFrame.fromJS(timeFrame));
     }
 
     case ExpressionType.Number: {
-      const { preparedValue, value } = (pojo as any as ITermExpression<number>);
-      return new NumberExpression(value, preparedValue);
+      const { value } = (pojo as any as ITermExpression);
+      return new NumberExpression(value);
     }
 
     case ExpressionType.Text: {
-      const { preparedValue, value } = (pojo as any as ITermExpression<string>);
-      return new TextExpression(value, preparedValue);
+      const { value } = (pojo as any as ITermExpression);
+      return new TextExpression(value);
     }
 
     case ExpressionType.Selector: {
-      const { matchingType, preparedValue, value } = (pojo as any as ISelectorExpression);
-      return new SelectorExpression(matchingType, value, preparedValue);
+      const { matchingType, value } = (pojo as any as ISelectorExpression);
+      return new SelectorExpression(matchingType, value);
     }
 
     case ExpressionType.Phrase: {
-      const { term } = (pojo as any as IPhraseExpression<any>);
+      const { term } = (pojo as any as IPhraseExpression);
       return new PhraseExpression(fromJS(term) as TermExpression);
     }