+++ src/common/utils/index.ts
@@ -1,1 +1,5 @@
+export * from './is-date';
+export * from './is-number';
 export * from './ordinal';
+export * from './parse-date';
+export * from './parse-number';