+++ src/common/utils/parse-date.ts
@@ -1,0 +1,6 @@
+import { parse } from 'chrono-node';
+
+import { TimeFrame } from '../model';
+
+// TODO: Drop wrong parsing results
+export const parseDate = (value: string): TimeFrame => TimeFrame.fromChrono(parse(value), value);