+++ src/dto/expressions/index.ts
@@ -1,4 +1,5 @@
 export * from './binary';
+export * from './date';
 export * from './expression';
 export * from './function';
 export * from './not';