+++ src/dto/expressions/selector.ts
@@ -1,6 +1,6 @@
 import { ValueType } from '../../common/model';
 import { ITermExpression } from './term';
 
-export interface ISelectorExpression extends ITermExpression<string> {
+export interface ISelectorExpression extends ITermExpression {
   readonly matchingType: ValueType;
 }